refactor : 명예의전당 리팩토링
