+++ src/domain/Honor/HonorDetail/index.tsx
@@ -1,4 +1,4 @@
-import { Layout } from '../../../components/common/Layout';
+import { Layout } from 'components/common/Layout';
 
 export const HonorDetail = () => {
   return <Layout>HonorDetail</Layout>;
