+++ src/domain/Honor/HonorList/List/List.style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Container = styled.ul`
+  display: flex;
+  gap: 50px;
+  flex-wrap: wrap;
+`;
+
+export const ItemContainer = styled.li``;
