+++ src/domain/Honor/HonorList/styles.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  margin: 40px 0;
+`;
+
+export const Title = styled.h2`
+  font-size: 2rem;
+  font-weight: bold;
+  margin-bottom: 40px;
+`;
